requirements.txt
smolagents
python-dotenv
duckduckgo_search
litellm
streamlit